Factor out js import/export code.

+++ octant-gui-core/src/pub_key_cred_params.rs
@@ -4,5 +4,5 @@ use serde::{Deserialize, Serialize};
 pub struct PubKeyCredParams {
     #[serde(rename = "type")]
     pub typ: String,
-    pub alg: i64,
+    pub alg: i32,
 }
